[HttpKernel] Fix deprecation of Extension::addClassesToCompile() / AddClassesToCachePass
