+++ UPGRADE-4.0.md
@@ -364,6 +364,8 @@ HttpFoundation
 HttpKernel
 ----------
 
+ * The `Extension::addClassesToCompile()` method has been removed.
+
  * Possibility to pass non-scalar values as URI attributes to the ESI and SSI
    renderers has been removed. The inline fragment renderer should be used with
    non-scalar attributes.
