+++ src/Symfony/Component/HttpKernel/CHANGELOG.md
@@ -4,12 +4,13 @@ CHANGELOG
 3.3.0
 -----
 
- * Added `kernel.project_dir` and `Kernel::getProjectDir()`
- * Deprecated `kernel.root_dir` and `Kernel::getRootDir()`
- * Deprecated `Kernel::getEnvParameters()`
- * Deprecated the special `SYMFONY__` environment variables
+ * added `kernel.project_dir` and `Kernel::getProjectDir()`
+ * deprecated `kernel.root_dir` and `Kernel::getRootDir()`
+ * deprecated `Kernel::getEnvParameters()`
+ * deprecated the special `SYMFONY__` environment variables
  * added the possibility to change the query string parameter used by `UriSigner`
  * deprecated `LazyLoadingFragmentHandler::addRendererService()`
+ * deprecated `Extension::addClassesToCompile()`
 
 3.2.0
 -----
